Manion_ver1.1.1/pyproject.toml
[tool.poetry]
name = "manimator"
version = "0.1.0"
description = "Transform research papers and mathematical concepts into stunning visual explanations (using the manim engine), powered by AI"
authors = ["Samarth P <hypercluster.tech@gmail.com>", "Vyoman Jain <hypercluster.tech@gmail.com>", "Shiva Golugula <hypercluster.tech@gmail.com>", "Sai Sathvik <hypercluster.tech@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "manimator"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
litellm = "^1.56.10"
python-multipart = "^0.0.20"
tenacity = "^9.0.0"
manim = "^0.18.1"
pypdf2 = "^3.0.1"
gradio = "^5.9.1"
google-cloud-vision = "^3.4.5"
pillow = "^10.0.0"
requests = "^2.31.0"
pdf2image = "^1.17.0"
poppler-utils = "^0.1.0"

[tool.poetry.scripts]
app = "manimator.main:main"
gradio-app = "manimator.gradio_app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
